[Fix] SignUpVC extension 분리 가독성 높이기 위해 SignUpVC에 작성되어 있던 extension 분리
